statistics: add tests for lock/unlock stats
