+++ statistics/handle/lockstats/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "lockstats",
     srcs = [
         "lock_stats.go",
+        "query_lock.go",
         "unlock_stats.go",
     ],
     importpath = "github.com/pingcap/tidb/statistics/handle/lockstats",
